moves all cryptokeys code to passphrase.go
